+++ R/globals.R
@@ -1,4 +1,4 @@
-utils::globalVariables(c("df_fut","dflong","dfwide","expiry_table","holidaysOil",
+utils::globalVariables(c("df_fut","dflong","dfwide","expiry_table","holidaysOil","EIAkey",
                          "a","x","df","value","series","symbol",".","price",
                          "cmdty","Date","expiry","V1","yield","maturity","contract","us.df","cf",
                          "DoW","Last.Trade","Year","key","swap","tick.prefix","Month.Letter",