+++ NEWS.md
@@ -1,3 +1,7 @@
+# RTL 0.1.3
+
++ `eia2tidy` removed dependency to `EIAdata` package.
+
 # RTL 0.1.2
 
 + `expiry_table` fixed for 2021-2024 CME WTI Futures.